requirements.txt
blinker
deepface
mediapipe
opencv-python
numpy
tensorflow
pillow
mysql-connector-python
cvzone